+++ packages/next/src/server/web/sandbox/context.ts
@@ -28,11 +28,11 @@ interface ModuleContext {
   warnedEvals: Set<string>
 }
 
-let getServerError: typeof import('../../../client/components/react-dev-overlay/pages/middleware').getServerError
-let decorateServerError: typeof import('../../../client/components/react-dev-overlay/pages/middleware').decorateServerError
+let getServerError: typeof import('../../../client/components/react-dev-overlay/server/middleware').getServerError
+let decorateServerError: typeof import('../../../client/components/react-dev-overlay/server/middleware').decorateServerError
 
 if (process.env.NODE_ENV === 'development') {
-  const middleware = require('../../../client/components/react-dev-overlay/pages/middleware')
+  const middleware = require('../../../client/components/react-dev-overlay/server/middleware')
   getServerError = middleware.getServerError
   decorateServerError = middleware.decorateServerError
 } else {
